// front-end/src/types/item.ts
export type Item = {
  id: number
  title: string
  description: string
  category_id: number
  photo: string
  price: number
  caution: number
  owner_id: number
  is_available: boolean
  created_at: string

   user_user_type: string;
  user_city: string;
  user_rating_avg: number;
  user_rating_count: number;
  user_first_name: string;
  user_last_name: string;
  user_photo: string;
}